pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["rxplus*"]
exclude = ["resources*"]

[project]
name = "rxplus"
version = "0.1.6"
description = "A package for reactive programming extensions."
readme = "README.md"
authors = [
    {name = "Yingte", email = "lucianoxu@foxmail.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.7"

# Basic dependencies only - core reactive programming utilities
dependencies = [
    "reactivex>=4.0.4",
    "websockets>=15.0.1",
    "prompt-toolkit>=3.0.51",
    "attrs>=25.4.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
]

[project.optional-dependencies]
# Audio processing support (microphone, speaker, wav files)
audio = [
    "pyaudio>=0.2.14",
    "soundfile>=0.12.1",
    "scipy>=1.13.1",
    "numpy",
]

# Video/graphics processing support (screen capture, image conversion)
video = [
    "pillow>=11.3.0",
    "mss>=10.0.0",
    "numpy",
]

# Install all optional dependencies
all = [
    "rxplus[audio,video]",
]

# Development dependencies
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.0",
    "mypy>=1.19.0",
    "pre-commit>=4.0.0",
]

[tool.ruff]
target-version = "py313"
line-length = 88
extend-exclude = ["ai-instru"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
]

[tool.mypy]
python_version = "3.13"
warn_unused_configs = true
disallow_any_generics = false
disallow_untyped_defs = false
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
no_implicit_optional = true

[[tool.mypy.overrides]]
module = [
    "reactivex.*",
    "mss.*",
    "pyaudio.*",
    "soundfile.*",
    "scipy.*",
    "numpy.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=rxplus --cov-report=term-missing"